Silence compiler warnings and adopt updated project settings
Type id-typed members against DataSourceProtocol so ambiguous `kind`
and `viewID` selectors resolve unambiguously, add explicit casts for
shorten-64-to-32 truncations across the codebase, drop empty Carbon
Rez build phase, and turn off three Xcode-recommended settings that
don't fit this informal-protocol codebase (user script sandboxing,
undeclared-selector, deprecated-implementations).

+++ Sources/AXAUComponent.m
@@ -281,7 +281,7 @@ OSStatus inputCallback(void     *inRefCon,
     for (c = 0; c < theNumberOfChannels; c++) 
     {
         theAudioData->mBuffers[c].mNumberChannels = 1;
-        theAudioData->mBuffers[c].mDataByteSize = theNumberOfFramesPerSlice * sizeof(Float32);
+        theAudioData->mBuffers[c].mDataByteSize = (UInt32)(theNumberOfFramesPerSlice * sizeof(Float32));
     }
     
     long slice;
@@ -292,12 +292,12 @@ OSStatus inputCallback(void     *inRefCon,
         for (c = 0; c < theNumberOfChannels; c++)
             theAudioData->mBuffers[c].mData = NULL;
 
-        AudioUnitRender (mAudioUnit, 
-                            &theActionFlags, 
-                            &theTimeStamp, 
-                            0, 
-                            theNumberOfFramesPerSlice, 
-                            theAudioData); 
+        AudioUnitRender (mAudioUnit,
+                            &theActionFlags,
+                            &theTimeStamp,
+                            0,
+                            (UInt32)theNumberOfFramesPerSlice,
+                            theAudioData);
         
         for (c = 0; c < theNumberOfChannels; c++)
         {
